add description of project improved

+++ Makefile
@@ -3,5 +3,5 @@ documentation:
 	doxygen doxy_config
 	if [ ! -d "doc/html/images" ]; then ln -s ../../images doc/html/images; fi
 markdown:
-	rm README.html
+	if [ -e "README.html" ]; then rm README.html; fi
 	markdown README.md >> README.html
